requirements.txt
aes==1.2.0
asgiref==3.8.1
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fsspec==2024.3.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.22.2
idna==3.6
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.42.0
MarkupSafe==2.1.3
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
openai-whisper==20231117
opencc-python-reimplemented==0.1.7
packaging==24.0
pillow==10.2.0
pycparser==2.22
pycryptodome==3.20.0
PyMySQL==1.1.0
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
simple-websocket==1.0.0
SQLAlchemy==2.0.29
sympy==1.12
tbb==2021.13.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.3.1
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.8.0
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
